Add guide, backfill options (initial partition)
Still writing the end of this, and will pare it down both for the
stack guide and elsewhere, but the main extension repo will hold the
full version (best though of as a use-case how-to).

Because the demo use case requires backfilling data, I've added a
parameter to enable_ts_config to request arbitrarily early partition
creation.