Bump to 0.1.2, add backfill, guide, PG fixes
This adds a new parameter to `enable_ts_table` which permits easy creation of partitions back to any arbitrary point in the past; great for backfills.

In addition, the use of a PostgreSQL 16-only syntax in the extension script was breaking compatibility with PostgreSQL 14 and 15. This has been removed.

Finally, there is also a big guide now.